working python 3 version
